dont stop execution on transaction errors
